+++ src/main/java/com/gbdevteam/teamnotes/controller/NoteController.java
@@ -1,54 +1,61 @@
 package com.gbdevteam.teamnotes.controller;
 
-import com.gbdevteam.teamnotes.model.Note;
+import com.gbdevteam.teamnotes.dto.NoteDTO;
+import com.gbdevteam.teamnotes.model.BoardRoleEnum;
+import com.gbdevteam.teamnotes.service.BoardRoleService;
 import com.gbdevteam.teamnotes.service.NoteService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.security.Principal;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 import java.util.UUID;
 
 @RequestMapping("/api/v1/note")
 @RestController
 @RequiredArgsConstructor
 @Slf4j
-public class NoteController {
+@Validated
 
+public class NoteController {
     private final NoteService noteService;
+    private final BoardRoleService boardRoleService;
 
     @GetMapping
-    public List<Note> findAll(){
+    public List<NoteDTO> findAll() {
         return noteService.findAll(UUID.randomUUID());
     }
 
     @GetMapping("/{id}")
-    public Optional<Note> getOneNotedById(@PathVariable UUID id) throws Throwable {
-        noteService.findById(id).orElseThrow(() ->
-                new NoSuchElementException(
-                        "Note with id:" + id + "doesn't exist"));
+    public NoteDTO getOneNotedById(@PathVariable("id") UUID id) {
         return noteService.findById(id);
     }
 
     @PostMapping
-    public void create(@RequestBody Note note){
-        noteService.create(note);
+    public void create(@Valid @RequestBody NoteDTO note, Principal principal) {
+        if (!boardRoleService.checkRole(note.getBoardId(), principal.getName()).equals(BoardRoleEnum.READER)) {
+            log.info(note.toString());
+            noteService.create(note);
+        }
     }
 
-    //example, not implemented
     @PutMapping
-    public void update(@RequestBody Note note){
-        noteService.update(note);
+    public void update(@Valid @RequestBody NoteDTO note, Principal principal) {
+        if (!boardRoleService.checkRole(note.getBoardId(), principal.getName()).equals(BoardRoleEnum.READER)) {
+            noteService.update(note);
+        }
     }
 
-    //example, not implemented
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable UUID id){
-        noteService.deleteById(id);
+    public void deleteById(@PathVariable("id") UUID id, Principal principal) {
+        if (!boardRoleService.checkRole(noteService.findById(id).getBoardId(), principal.getName()).equals(BoardRoleEnum.READER)) {
+            noteService.deleteById(id);
+        }
     }
 
 }
